+++ flmodules/src/lib.rs
@@ -1,6 +1,8 @@
+pub mod broker;
+pub mod nodeids;
+
+pub mod template;
 pub mod random_connections;
 pub mod gossip_events;
 pub mod timer;
 pub mod ping;
-pub mod broker;
-pub mod nodeids;