+++ flmodules/src/random_connections/mod.rs
@@ -1,3 +1,3 @@
-pub mod module;
+pub mod messages;
 pub mod nodes;
 pub mod storage;